Add code to program creator

+++ app/src/main/java/com/bartoszdrozd/fitapp/model/program/ProgramValues.kt
@@ -5,5 +5,8 @@ import com.bartoszdrozd.fitapp.model.workout.OneRepMax
 data class ProgramValues(
     val programType: ProgramType,
     val maxes: List<OneRepMax>,
-    val trainingMax: Int
+    /**
+     * Value between 0-1
+     */
+    val trainingMax: Float
 )
